chore: Add ProductController and ProductRepository

+++ src/main/java/com/java/products/models/Product.java
@@ -11,7 +11,7 @@ public class Product {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id; 
+    private Long id; 
 
     @Column(nullable = false, length = 50)
     private String name;
@@ -22,13 +22,13 @@ public class Product {
     public Product() {
     }
 
-    public Product(int id, String name, double price) {
+    public Product(Long id, String name, double price) {
         this.id = id;
         this.name = name;
         this.price = price;
     }
 
-    public int getId() {
+    public Long getId() {
         return id;
     }
 
@@ -40,7 +40,7 @@ public class Product {
         return price;
     }
 
-    public void setId(int id) {
+    public void setId(Long id) {
         this.id = id;
     }
